Share render code between static and interactive diagrams
